Reduce MAX_READ_LENGTH from 512 to 16
With MAX_READ_LENGTH = 512, lot of data can be potentially received from
the serial port which would correspond to a lot of serialized packets
being received together. We don't want to wait for a bunch of packets
being received before publishing which adds latency to the measurements
so reduce the MAX_READ_LENGTH such that the read_callback is called
sooner after the data is received.